chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/animation-shorthand-expected.txt

This is a testharness.js-based test.
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim" should set the property value
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim timeline" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim none" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim \"initial\"" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim auto" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim initial" should not set the property value
PASS e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim 2000" should not set the property value
PASS Property animation value '1s linear 1s 2 reverse forwards paused anim'
FAIL Property animation value '1s linear 1s 2 reverse forwards paused anim timeline' assert_true: '1s linear 1s 2 reverse forwards paused anim timeline' is a supported value for animation. expected true got false
FAIL Property animation value '1s linear 1s 2 reverse forwards paused anim none' assert_true: '1s linear 1s 2 reverse forwards paused anim none' is a supported value for animation. expected true got false
FAIL Property animation value '1s linear 1s 2 reverse forwards paused anim auto' assert_true: '1s linear 1s 2 reverse forwards paused anim auto' is a supported value for animation. expected true got false
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-delay assert_equals: animation-delay should be canonical expected "1s, 1s, 1s" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-direction assert_equals: animation-direction should be canonical expected "reverse, reverse, reverse" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-duration assert_equals: animation-duration should be canonical expected "1s, 1s, 1s" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-fill-mode assert_equals: animation-fill-mode should be canonical expected "forwards, forwards, forwards" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-iteration-count assert_equals: animation-iteration-count should be canonical expected "2, 2, 2" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-name assert_equals: animation-name should be canonical expected "anim1, anim2, anim3" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-play-state assert_equals: animation-play-state should be canonical expected "paused, paused, paused" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-timeline assert_equals: animation-timeline should be canonical expected (string) "timeline, none, auto" but got (undefined) undefined
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should set animation-timing-function assert_equals: animation-timing-function should be canonical expected "linear, linear, linear" but got ""
FAIL e.style['animation'] = "1s linear 1s 2 reverse forwards paused anim1 timeline,\n   1s linear 1s 2 reverse forwards paused anim2 none,\n   1s linear 1s 2 reverse forwards paused anim3 auto" should not set unrelated longhands assert_true: expected true got false
Harness: the test ran to completion.
